Update leagueoflegends.json
Fixes a problem with the game creating a new dir under the main one.

With the current manifest, the game creates a new "Riot Games" directory at "scoop\apps\APP", one level under the installation dir. The fix is to resemble the default installation path which is "C:\Riot Games\League of Legends".